Title: 알파벳과 사칙연산, Time: 1275ms, Memory: 176MB, Status: Memory Limit Exceed - CodeTree

+++ 241009/알파벳과 사칙연산/calculations-with-alphabet.java
@@ -9,7 +9,7 @@ public class Main {
     static int res[];
     static int arr[] ={1,2,3,4};
     static int n;
-    static int maxVal = Integer.MIN_VALUE;
+    static long maxVal = Integer.MIN_VALUE;
     static String str[];
 
     public static void main(String[] args) throws IOException {
@@ -22,7 +22,7 @@ public class Main {
     static void permutation(int cnt) {
         if(cnt == n) {
             //res의 결과
-            int c = calc();
+            long c = calc();
             maxVal = Math.max(maxVal, c);
             return;
         }
@@ -32,14 +32,14 @@ public class Main {
         }
     }
 
-    static int calc() {
+    static long calc() {
         int idx=0;
         for(int i=0; i< str.length; i+=2) { //짝수만
             str[i] = String.valueOf(res[idx]);
             idx++;
         }
         //str 가지고 계산하기
-        int sum = Integer.parseInt(str[0]);
+        long sum = Integer.parseInt(str[0]);
         for(int i=1; i<str.length-1; i++) {
             if(i%2==1) { //연산자일 때
                 if(str[i].equals("-")) {
